improved cartRoute

+++ utils/routes/cartRoute.js
@@ -17,20 +17,35 @@ routerCart.post('/', (req, res) => {
 
 routerCart.delete('/:id', (req, res) => {
     const thisCartId = parseFloat(req.params.id);
-    try{
-        const deletedItem = cart.getById(thisCartId);
-        cart.deleteById(thisCartId);
-        res.status(200).send(JSON.stringify({deletedItem}));
-    } catch (e) {
-        res.status(500).send(JSON.stringify({error: "ID carrito inexistente"}));
-    }
+    cart.getById(thisCartId)
+    .then((thisCart) => {
+        if(thisCart){
+            cart.deleteById(thisCartId)
+            .then((deletedItem) => {
+                res.status(200).send(JSON.stringify({deletedItem}));
+            }).catch((err) => {
+                logger.error(`Error when deleting cart with id ${thisCartId}`);
+                res.status(500).json({at:'cartDelete', error: err.message});
+            })
+        } else {
+            logger.error(`Invalid id sent to delete cart ${thisCartId}`);
+            res.status(400).json({at:'cartDelete', error: `Invalid id sent to delete cart ${thisCartId}`});
+        }
+    })
+    .catch((err) => {
+        logger.error(`Error finding cart to delete with id ${thisCartId}`);
+        res.status(400).json({at:'cartDelete', error: err.message});
+    })
 })
 
 routerCart.get('/:id/productos', (req, res) => {
     const thisCartId = parseFloat(req.params.id);
     cart.getById(thisCartId)
-    .then((thisCart) => res.status(200).send(JSON.stringify(thisCart.productos)))
-    .catch((err) => res.status(500).send(JSON.stringify({error: "ID carrito inexistente"})))
+    .then((thisCart) => res.status(200).json(thisCart.productos))
+    .catch((err) => {
+        logger.error(`Error searching for cart products, ${err.message}`)
+        res.status(500).send(JSON.stringify({error: "ID carrito inexistente"}))
+    })
 })
 
 routerCart.post('/:id/productos/:id_prod', (req, res) => {
@@ -52,16 +67,16 @@ routerCart.post('/:id/productos/:id_prod', (req, res) => {
             cart.change(thisCartId, thisCart)
             .then((changedCart) => res.status(200).send(JSON.stringify(changedCart)))
             .catch((e) => {
-                logger.error(e.message)
-                res.status(500).send(JSON.stringify({error: "Error al actualizar carrito"}))
+                logger.error("Invalid product id or " + err.message)
+                res.status(500).send(JSON.stringify({at: "cartUpdate", error: "ID producto inexistente o " + err.message}))
             })
         }).catch((e) => {
-            logger.error(e.message)
-            res.status(500).send(JSON.stringify({error: "ID producto inexistente"}))
+            logger.error("Invalid product id or " + err.message)
+            res.status(500).send(JSON.stringify({at: "cartUpdate", error: "ID producto inexistente o " + err.message}));
     })
     }).catch((e) => {
-        logger.error(e.message)
-        res.status(500).send(JSON.stringify({error: "ID carrito inexistente"}))
+        logger.error("Invalid cart id or " + err.message)
+        res.status(500).send(JSON.stringify({error: "ID carrito inexistente o " + err.message}))
     })
 })
 
@@ -87,19 +102,18 @@ routerCart.delete('/:id/productos/:id_prod', (req, res) => {
                 cart.change(thisCartId, thisCart);
                 res.status(200).send(JSON.stringify(thisCart));
             } else {
-                res.status(200).send("There's no product of that ID in the requested Cart\n" + JSON.stringify(thisCart))
+                res.status(400).send("No existe un producto con ese ID\n" + JSON.stringify(thisCart))
             }
         })
         .catch((err) => {
-            logger.error(err.message);
-            res.status(500).send(JSON.stringify({error: "ID producto inexistente"}));
+            logger.error("Invalid product ID or " +err.message);
+            res.status(500).send(JSON.stringify({at: "deleteProdFromCart", error: "ID producto inexistente o " + err.message}));
         })
     })
     .catch((err) => {
-        logger.error(err.message);
-        res.status(500).send(JSON.stringify({error: "ID carrito inexistente"}));
+        logger.error("Invalid cart ID or " +err.message);
+        res.status(500).send(JSON.stringify({at: "deleteProdFromCart", error: "ID carrito inexistente o " + err.message}));
     })
-        
 })
 
 export default routerCart
